move secure DNSSEC scenarios into a module
